Add files via upload
These changes in setPosition method in Car class is made to solve the bug about dragging the cars. Before the changes we could drag the cars to space even there is a car between the target space and the current position of the car but now, we cannot do it.