Add files via upload
The changes in Car class is made to solve a bug. Before the changes, we cannot drag the car on another car but we could drag the car an available space even there is a car between the target space and the current area of the car. Now we cannot do it and there is no bug.